requirements.txt
pycryptodome
numpy
Pillow
opencv-python
streamlit
